Improves game stats tracking
Updates the game stats system to include more comprehensive data such as total damage, headshot kills, first kills, first deaths, clutch attempts, and clutch wins.

Modifies the 'ClutchRatio' property to be a double to represent the ratio more accurately.

Calculates the clutch ratio based on clutch wins and attempts.

Adds new fields to the database to persist the new stats.

+++ src/Application/GameStats/Create/CreateGameStatCommand.cs
@@ -10,7 +10,7 @@ public sealed class CreateGameStatCommand : ICommand<Guid>
     public int Deaths { get; set; }
     public int Adr { get; set; }
     public int HsRatio { get; set; }
-    public int ClutchRatio { get; set; }
+    public double ClutchRatio { get; set; }
     public int Fkpr { get; set; }
     public int Fdpr { get; set; }
     public int RoundsPlayed { get; set; }
